Final optimizations(?)
Complete rewrite of the analysis code; instead of mixing the data with transpose() then calling ioc() on each chunk, determine the correct indices as we're counting characters for ioc(). Additionally improve frequency analysis by using the same Vec<> to count all the characters, and clean up non-zero indices once we're finished with it. This results in a ~4x speedup, which is probably as fast as it's going to get single-threaded.